Fixed page 4
Needed to be a double declaration, not int to calculate the averages.

+++ lab/Lab - Functions.playground/Pages/4. App Exercise - Progress Updates.xcplaygroundpage/Contents.swift
@@ -3,7 +3,7 @@
  
  In many cases you want to provide input to a function. For example, the report grade function you wrote page 2 might be located in an area of your project that doesn't have access to the value of `correctQuestions` and `totalQuestions`. In that case, whenever you called the function, you would need to provide it with that information so it can print the correct grade.
  
- Rewrite the function `reportGrade`, only this time give it two parameters of type `Int` called `totalQuestions` and `correctQuestions`, respectively. It should work exactly the same way as before using the values passed. Call the function and observe the printout.
+ Rewrite the function `reportGrade`, only this time give it two parameters of type `Double` called `totalQuestions` and `correctQuestions`, respectively. It should work exactly the same way as before using the values passed. Call the function and observe the printout.
  
  **Call the function a number of times**, passing in different values of `totalQuestions` and `correctQuestions`. Observe the printouts and make sure what is printed to the console is what you would expect for the parameters passed in. (1 point)
  */
